+++ packages/blinkid-verify/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @microblink/blinkid-verify
 
+## 3.20.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @microblink/blinkid-verify-core@3.20.2
+  - @microblink/blinkid-verify-ux-manager@3.20.2
+
 ## 3.20.1
 
 ### Patch Changes